cli: refactor shell code into multiple files
